+++ app/controllers/accounts_controller.rb
@@ -1,23 +1,24 @@
+# frozen_string_literal: true
+
 class AccountsController < ApplicationController
   before_action :set_user
-  authorize_resource :user 
+  authorize_resource :user
 
-  def edit 
-  end
+  def edit; end
 
-  def update 
+  def update
     @user.update(user_params)
     respond_with @user, location: edit_account_path
   end
 
-  private 
+  private
 
-  def set_user 
+  def set_user
     @user = current_user
-    raise CanCan::AccessDenied.new unless @user.present?
+    raise CanCan::AccessDenied if @user.blank?
   end
 
-  def user_params 
+  def user_params
     params.require(:user).permit(:password, :password_confirmation)
   end
 end